requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==22.12.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
coverage==7.0.1
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.0
executing==1.2.0
flake8==5.0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
hydra-core==1.3.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.2
iniconfig==1.1.1
ipython==8.7.0
isort==5.11.4
jedi==0.18.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
packaging==22.0
parso==0.8.3
pathspec==0.10.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
pluggy==1.0.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenize-rt==5.0.0
tomli==2.0.1
torch==1.8.1+cu111
torchinfo==1.7.1
torchmetrics==0.11.0
torchtyping==0.1.4
torchvision==0.9.1
tqdm==4.64.1
traitlets==5.8.0
typeguard==2.13.3
types-tqdm==4.64.7.9
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.7
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
